Move `class FrontlineParser` into `frontline_parser.cpp`
`ExtractCvdArgs` is the only public interface actually used.

Bug: b/389804379
Test: /usr/bin/bazel run :cvd -- --instance_name=1 status
