+++ base/cvd/cuttlefish/host/commands/cvd/cli/frontline_parser.cpp
@@ -17,33 +17,66 @@
 #include "host/commands/cvd/cli/frontline_parser.h"
 
 #include <memory>
+#include <optional>
+#include <string>
 #include <vector>
 
 #include <android-base/file.h>
 #include <android-base/strings.h>
 
 #include "host/commands/cvd/cli/flag.h"
+#include "host/commands/cvd/cli/selector/arguments_separator.h"
 #include "host/commands/cvd/cli/selector/selector_constants.h"
 #include "host/commands/cvd/cli/types.h"
 
 namespace cuttlefish {
+namespace {
 
-Result<cvd_common::Args> ExtractCvdArgs(cvd_common::Args& args) {
-  auto frontline_parser = CF_EXPECT(FrontlineParser::Parse(args));
-  CF_EXPECT(frontline_parser != nullptr);
+/* the very first command line parser
+ *
+ * Being aware of valid subcommands and cvd-specific commands, it will
+ * separate the command line arguments into:
+ *
+ *  1. program path/name
+ *  2. cvd-specific arguments
+ *     a) selector flags
+ *     b) non-selector flags
+ *  3. subcommand
+ *  4. subcommand arguments
+ *
+ * This is currently on the client side but will be moved to the server
+ * side.
+ */
+class FrontlineParser {
+  using ArgumentsSeparator = selector::ArgumentsSeparator;
 
-  const auto prog_path = frontline_parser->ProgPath();
-  const auto new_sub_cmd = frontline_parser->SubCmd();
-  cvd_common::Args cmd_args{frontline_parser->SubCmdArgs()};
+ public:
+  // This call must guarantee all public methods will be valid
+  static Result<std::unique_ptr<FrontlineParser>> Parse(
+      const cvd_common::Args& all_args);
 
-  cvd_common::Args new_exec_args{prog_path};
-  if (new_sub_cmd) {
-    new_exec_args.push_back(*new_sub_cmd);
-  }
-  new_exec_args.insert(new_exec_args.end(), cmd_args.begin(), cmd_args.end());
-  args = new_exec_args;
-  return frontline_parser->CvdArgs();
-}
+  const std::string& ProgPath() const;
+  std::optional<std::string> SubCmd() const;
+  const cvd_common::Args& SubCmdArgs() const;
+  const cvd_common::Args& CvdArgs() const;
+
+ private:
+  FrontlineParser(const cvd_common::Args& all_args);
+
+  // internal workers in order
+  Result<void> Separate();
+  Result<std::unique_ptr<ArgumentsSeparator>> CallSeparator();
+  struct FilterOutput {
+    bool clean;
+    bool help;
+    cvd_common::Args selector_args;
+  };
+  Result<FilterOutput> FilterNonSelectorArgs();
+
+  const cvd_common::Args all_args_;
+  const std::vector<std::string> internal_cmds_;
+  std::unique_ptr<ArgumentsSeparator> arguments_separator_;
+};
 
 Result<std::unique_ptr<FrontlineParser>> FrontlineParser::Parse(
     const cvd_common::Args& all_args) {
@@ -92,6 +125,25 @@ const cvd_common::Args& FrontlineParser::SubCmdArgs() const {
 
 const cvd_common::Args& FrontlineParser::CvdArgs() const {
   return arguments_separator_->CvdArgs();
+}
+
+}  // namespace
+
+Result<cvd_common::Args> ExtractCvdArgs(cvd_common::Args& args) {
+  auto frontline_parser = CF_EXPECT(FrontlineParser::Parse(args));
+  CF_EXPECT(frontline_parser != nullptr);
+
+  const auto prog_path = frontline_parser->ProgPath();
+  const auto new_sub_cmd = frontline_parser->SubCmd();
+  cvd_common::Args cmd_args{frontline_parser->SubCmdArgs()};
+
+  cvd_common::Args new_exec_args{prog_path};
+  if (new_sub_cmd) {
+    new_exec_args.push_back(*new_sub_cmd);
+  }
+  new_exec_args.insert(new_exec_args.end(), cmd_args.begin(), cmd_args.end());
+  args = new_exec_args;
+  return frontline_parser->CvdArgs();
 }
 
 }  // namespace cuttlefish
